More work on echo
Still no attempt to make it compile. I added some documentation to the
sound code, though, and added code for commands F5-F7 (no F8 yet).

There's a lag between setting the echo delay from the player and seeing
that delay actually take effect in N-SPC, so I figured this was a good
opportunity to start creating a difference between the values that the
player code sees and the actual DSP registers the sound code sees.